# neomarket-canon/apis/b2c/cart/openapi.yaml
openapi: 3.0.3
info:
  title: "NeoMarket B2C: Корзина + Избранное + Главная"
  version: "0.1"
  description: |
    Проектируется синдикатом Interface.
    Домены: корзина (CRUD), избранное, главная страница (баннеры, подборки).

  contact:
    name: Interface Team
    email: interface@neomarket.ru
...
security:
- BearerAuth: []
paths:
...
  /api/v1/favorites:
    get:
      tags:
        - Избранное
      summary: Получение списка избранных товаров
      description: |
        Возвращает список товаров, добавленных авторизованным пользователем в избранное.

        **Авторизация:** требуется JWT-токен в заголовке `Authorization: Bearer <token>`.
        Идентификатор пользователя извлекается из токена автоматически.

        **Как это работает:**
        - Сервис хранит только `product_id` и дату добавления.
        - Для получения актуальной информации о товарах (названия, цены, наличие, SKU) выполняется batch-запрос к B2B-каталогу.
        - Товары, которые не найдены в B2B (например, удалены из каталога), **исключаются** из ответа (они автоматически удаляются или игнорируются). Клиент не получает информацию о таких товарах.
        - Если B2B временно недоступен — возвращается ошибка 503.

        **Пагинация:** поддерживается через параметры `limit` и `offset`.
        Поле `total` содержит общее количество избранных товаров (только доступных) без учёта пагинации.

        **Пустое избранное:** возвращается `items: []`, `total: 0`.
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          description: Максимальное количество элементов в ответе (пагинация)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          description: Смещение для пагинации
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Успешное получение списка избранного
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoritesResponse"
              examples:
                with_items:
                  summary: Список с товарами
                  value:
                    items:
                      - product:
                          id: "550e8400-e29b-41d4-a716-446655440000"
                          title: "iPhone 15 Pro Max"
                          description: "Флагман Apple..."
                          status: "MODERATED"
                          category:
                            id: "6ba7b810-9dad-11d1-80b4-00c04fd430c8"
                            name: "Смартфоны"
                          images:
                            - url: "/s3/iphone15-1.jpg"
                              ordering: 0
                          characteristics:
                            - name: "Бренд"
                              value: "Apple"
                          skus:
                            - id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                              name: "256GB Black"
                              price: 12999000
                              active_quantity: 5
                              characteristics:
                                - name: "Цвет"
                                  value: "Чёрный"
                                - name: "Объём памяти"
                                  value: "256 ГБ"
                            - id: "8a4e3f9c-1a2b-4c8d-9e5f-6b7a8c9d0e1f"
                              name: "256GB White"
                              price: 12999000
                              active_quantity: 0
                        added_at: "2025-03-09T12:00:00Z"
                    total: 1
                empty:
                  summary: Пустое избранное
                  value:
                    items: []
                    total: 0
        "400":
          description: Невалидные параметры запроса (например, limit/offset вне диапазона)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: "INVALID_PARAMETER"
                message: "Параметр limit должен быть от 1 до 100"
        "401":
          description: Пользователь не авторизован — отсутствует или недействителен JWT-токен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: "UNAUTHORIZED"
                message: "Требуется авторизация"
        "503":
          description: B2B сервис временно недоступен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: "B2B_UNAVAILABLE"
                message: "Сервис товаров временно недоступен, попробуйте позже"
  /api/v1/favorites/{product_id}:
    post:
      tags:
        - Избранное
      summary: Добавление товара в избранное
      description: |
        Добавляет товар в избранное пользователя.
        Идентификатор пользователя извлекается только из JWT claims.
        Если товар уже находится в избранном, можно вернуть 200,
        чтобы операция была безопасной для повторного вызова.
      security:
        - BearerAuth: []

      parameters:
        - name: product_id
          in: path
          required: true
          description: Идентификатор товара (UUID)
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"

      responses:
        "201":
          description: Товар успешно добавлен в избранное
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteMutationResponse"
              example:
                product_id: "550e8400-e29b-41d4-a716-446655440000"
                user_id: "123e4567-e89b-12d3-a456-426614174000"
                added_at: "2025-03-10T10:00:00Z"
                message: "Товар добавлен в избранное"

        "200":
          description: Товар уже был в избранном
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteMutationResponse"
              example:
                product_id: "550e8400-e29b-41d4-a716-446655440000"
                user_id: "123e4567-e89b-12d3-a456-426614174000"
                added_at: "2025-03-10T10:00:00Z"
                message: "Товар уже находится в избранном"

        "400":
          description: Невалидные параметры запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: "INVALID_PARAMETER"
                message: "Некорректный UUID product_id"

        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: "UNAUTHORIZED"
                message: "Требуется авторизация"

        "404":
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: "PRODUCT_NOT_FOUND"
                message: "Товар не найден"

        "503":
          description: B2B сервис временно недоступен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: "B2B_UNAVAILABLE"
                message: "Сервис товаров временно недоступен, попробуйте позже"

    delete:
      tags:
        - Избранное
      summary: Удаление товара из избранного
      description: |
        Удаляет товар из избранного пользователя.
        Идентификатор пользователя извлекается только из JWT claims.
        Если товара не было в избранном, можно всё равно вернуть 204.
      security:
        - BearerAuth: []

      parameters:
        - name: product_id
          in: path
          required: true
          description: Идентификатор товара (UUID)
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"

      responses:
        "204":
          description: Товар успешно удалён из избранного

        "400":
          description: Невалидные параметры запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: "INVALID_PARAMETER"
                message: "Некорректный UUID product_id"

        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: "UNAUTHORIZED"
                message: "Требуется авторизация"

        "404":
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: "PRODUCT_NOT_FOUND"
                message: "Товар не найден"

        "503":
          description: B2B сервис временно недоступен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: "B2B_UNAVAILABLE"
                message: "Сервис товаров временно недоступен, попробуйте позже"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT передаётся в заголовке Authorization: Bearer <token>.
        API Gateway проверяет подпись токена, извлекает userId
        и передаёт его в Cart Service через внутренний заголовок X-User-Id.
        Клиент не передаёт X-User-Id напрямую.
...
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "6ba7b810-9dad-11d1-80b4-00c04fd430c8"
        name:
          type: string
          example: "Смартфоны"
      required:
        - id
        - name
    Image:
      type: object
      properties:
        url:
          type: string
          example: "/s3/iphone15-1.jpg"
        ordering:
          type: integer
          example: 0
      required:
        - url
        - ordering
    Characteristic:
      type: object
      properties:
        name:
          type: string
          example: "Бренд"
        value:
          type: string
          example: "Apple"
      required:
        - name
        - value
    SKU:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
        name:
          type: string
          example: "256GB Black"
        price:
          type: integer
          description: Цена в копейках
          example: 12999000
        active_quantity:
          type: integer
          description: Доступное количество на складе
          example: 5
        characteristics:
          type: array
          items:
            $ref: "#/components/schemas/Characteristic"
      required:
        - id
        - name
        - price
        - active_quantity
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          example: "iPhone 15 Pro Max"
        description:
          type: string
          example: "Флагман Apple..."
        status:
          type: string
          enum: [CREATED, ON_MODERATION, MODERATED, BLOCKED]
          example: "MODERATED"
        category:
          $ref: "#/components/schemas/Category"
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"
        characteristics:
          type: array
          items:
            $ref: "#/components/schemas/Characteristic"
        skus:
          type: array
          items:
            $ref: "#/components/schemas/SKU"
      required:
        - id
        - title
        - description
        - status
        - category
        - images
        - characteristics
        - skus
    FavoriteItem:
      type: object
      properties:
        product:
          $ref: "#/components/schemas/Product"
        added_at:
          type: string
          format: date-time
          description: Дата и время добавления товара в избранное
          example: "2025-03-09T12:00:00Z"
      required:
        - product
        - added_at
    FavoritesResponse:
      type: object
      properties:
        items:
          type: array
          description: Список избранных товаров (обогащённых данными из B2B)
          items:
            $ref: "#/components/schemas/FavoriteItem"
        total:
          type: integer
          description: Общее количество товаров в избранном (без учёта пагинации)
          example: 5
      required:
        - items
        - total
    FavoriteMutationResponse:
      type: object
      properties:
        product_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        user_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        added_at:
          type: string
          format: date-time
          example: "2025-03-10T10:00:00Z"
        message:
          type: string
          example: "Товар добавлен в избранное"
      required:
        - product_id
        - user_id
        - added_at
        - message
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: "Код ошибки"
          example: "UNAUTHORIZED"
        message:
          type: string
          description: "Описание ошибки"
          example: "Требуется авторизация"
      required:
        - code
        - message

    # ----- СХЕМЫ ДЛЯ ПОДБОРОК продуктов -----
...